# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GYMCO customer API — multi-gym memberships, classes, bookings, check-ins, and fitness tracking.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assessment/book:
    post:
      operationId: bookAssessment
      tags: [trainers]
      summary: Book (or reschedule) the early-morning empty-stomach fitness assessment slot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookAssessmentRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssessmentBooking" }
...
components:
  securitySchemes:
    memberBearer:
      type: http
      scheme: bearer
    adminSession:
      type: apiKey
      in: cookie
      name: connect.sid
...
  schemas:
...
    AssessmentBooking:
      type: object
      required: [id, slotDate, slotTime, status, gymName, createdAt]
      properties:
        id: { type: integer }
        slotDate: { type: string, description: "YYYY-MM-DD (IST)" }
        slotTime: { type: string, description: "HH:MM 24h (IST), early morning" }
        status: { type: string, enum: [booked, completed, cancelled] }
        gymName: { type: string }
        createdAt: { type: string, format: date-time }
...
    BookAssessmentRequest:
      type: object
      required: [slotDate, slotTime]
      properties:
        slotDate: { type: string, description: "YYYY-MM-DD (IST); today or later" }
        slotTime: { type: string, description: "One of the bookable morning slot times" }
...